+++ panels/base_panel.py
@@ -4,7 +4,7 @@ import gi
 import logging
 
 gi.require_version("Gtk", "3.0")
-from gi.repository import Gtk, Gdk, GLib
+from gi.repository import Gtk, Gdk, GLib, Pango
 from jinja2 import Environment, Template
 
 from ks_includes.KlippyGtk import KlippyGtk
@@ -18,7 +18,8 @@ class BasePanel(ScreenPanel):
         self.time_min = -1
         self.time_format = self._config.get_main_config_option("24htime")
         self.title_spacing = self._gtk.font_size * 2
-
+        self.time_update = None
+        self.titlebar_name_type = None
         self.buttons_showing = {
             'back': False if back else True,
             'macros_shortcut': False,
@@ -88,13 +89,11 @@ class BasePanel(ScreenPanel):
             logging.debug("Error parsing jinja for title: %s" % title)
 
         self.titlelbl = Gtk.Label()
-        if self._screen.vertical_mode:
-            self.titlelbl.set_size_request(self._screen.width, self.title_spacing)
-        else:
-            self.titlelbl.set_size_request(self._screen.width - action_bar_width, self.title_spacing)
         self.titlelbl.set_hexpand(True)
+        self.titlelbl.set_vexpand(True)
         self.titlelbl.set_halign(Gtk.Align.CENTER)
-        self.titlelbl.set_valign(Gtk.Align.CENTER)
+        self.titlelbl.set_ellipsize(True)
+        self.titlelbl.set_ellipsize(Pango.EllipsizeMode.END)
         self.set_title(title)
 
         self.hmargin = 5
@@ -114,30 +113,32 @@ class BasePanel(ScreenPanel):
 
         self.control['time_box'] = Gtk.Box()
         self.control['time_box'].set_halign(Gtk.Align.END)
-        self.control['time_box'].set_size_request(0, self.title_spacing)
         self.control['time'] = Gtk.Label("00:00 AM")
-        self.control['time'].set_size_request(0, self.title_spacing)
-        self.control['time'].set_halign(Gtk.Align.END)
-        self.control['time'].set_valign(Gtk.Align.CENTER)
-        self.control['time_box'].pack_end(self.control['time'], True, 0, 0)
+        self.control['time_box'].pack_end(self.control['time'], True, True, self.hmargin)
 
         self.control['temp_box'] = Gtk.Box()
-        self.control['temp_box'].set_vexpand(True)
-        self.control['temp_box'].set_size_request(0, self.title_spacing)
+
+        self.titlebar = Gtk.Grid()
+        self.titlebar.set_valign(Gtk.Align.CENTER)
+        self.titlebar.set_property("column-spacing", 5)
+        if self._screen.vertical_mode:
+            self.titlebar.set_size_request(self._screen.width, self.title_spacing)
+        else:
+            self.titlebar.set_size_request(self._screen.width - action_bar_width, self.title_spacing)
+        self.titlebar.attach(self.control['temp_box'], 0, 0, 1, 1)
+        self.titlebar.attach(self.titlelbl, 1, 0, 1, 1)
+        self.titlebar.attach(self.control['time_box'], 2, 0, 1, 1)
 
         if self._screen.vertical_mode:
-            self.layout.put(self.control['temp_box'], 0, 0)
-            self.layout.put(self.titlelbl, 0, 0)
-            self.layout.put(self.control['time_box'], 0, 0)
+            self.layout.put(self.titlebar, 0, 0)
             self.layout.put(self.content, self.hmargin, self.title_spacing)
         else:
-            self.layout.put(self.control['temp_box'], action_bar_width, 0)
-            self.layout.put(self.titlelbl, action_bar_width, 0)
-            self.layout.put(self.control['time_box'], action_bar_width, 0)
+            self.layout.put(self.titlebar, action_bar_width, 0)
             self.layout.put(self.content, action_bar_width + self.hmargin, self.title_spacing)
 
+
     def initialize(self, panel_name):
-        # Create gtk items here
+        self.update_time()
         return
 
     def show_heaters(self, show=True):
@@ -147,34 +148,80 @@ class BasePanel(ScreenPanel):
         if show is False:
             return
 
+        for device in self._printer.get_temp_store_devices():
+            self.labels[device + '_box'] = Gtk.Box(spacing=0)
+            self.labels[device] = Gtk.Label(label="100º")
+            self.labels[device].set_ellipsize(True)
+            self.labels[device].set_ellipsize(Pango.EllipsizeMode.START)
+            if device.startswith("extruder"):
+                if device == "extruder":
+                    ext_img = self._gtk.Image("extruder-0.svg", None, .5, .5)
+                else:
+                    ext_img = self._gtk.Image("extruder-%s.svg" % device[8:], None, .5, .5)
+                self.labels[device + '_box'].pack_start(ext_img, True, True, 3)
+            elif device.startswith("heater_bed"):
+                bed_img = self._gtk.Image("bed.svg", None, .5, .5)
+                self.labels[device + '_box'].pack_start(bed_img, True, True, 3)
+            elif device.startswith("temperature_fan"):
+                fan_img = self._gtk.Image("fan.svg", None, .5, .5)
+                self.labels[device + '_box'].pack_start(fan_img, True, True, 3)
+            else:
+                temp_img = self._gtk.Image("heat-up.svg", None, .5, .5)
+                self.labels[device + '_box'].pack_start(temp_img, True, True, 3)
+            self.labels[device + '_box'].pack_start(self.labels[device], True, True, 0)
+
+        if self._screen.width <= 480:
+            nlimit = 3
+        elif self._screen.width <= 800:
+            nlimit = 4
+        else:
+            nlimit = 5
+
+        n = 0
         if self._printer.get_tools():
-            for i, extruder in enumerate(self._printer.get_tools()):
-                self.labels[extruder + '_box'] = Gtk.Box(spacing=0)
-                self.labels[extruder] = Gtk.Label(label="")
-                if i <= 4:
-                    ext_img = self._gtk.Image("extruder-%s.svg" % i, None, .4, .4)
-                    self.labels[extruder + '_box'].pack_start(ext_img, True, 3, 3)
-                self.labels[extruder + '_box'].pack_start(self.labels[extruder], True, 3, 3)
             self.current_extruder = self._printer.get_stat("toolhead", "extruder")
-            self.control['temp_box'].pack_start(self.labels["%s_box" % self.current_extruder], True, 5, 5)
+            self.control['temp_box'].pack_start(self.labels["%s_box" % self.current_extruder], True, True, 3)
+            n += 1
 
         if self._printer.has_heated_bed():
-            heater_bed = self._gtk.Image("bed.svg", None, .4, .4)
-            self.labels['heater_bed'] = Gtk.Label(label="20 C")
-            heater_bed_box = Gtk.Box(spacing=0)
-            heater_bed_box.pack_start(heater_bed, True, 5, 5)
-            heater_bed_box.pack_start(self.labels['heater_bed'], True, 3, 3)
-            self.control['temp_box'].pack_end(heater_bed_box, True, 3, 3)
+            self.control['temp_box'].pack_start(self.labels['heater_bed_box'], True, True, 3)
+            n += 1
 
+        # Options in the config have priority
+        printer_cfg = self._config.get_printer_config(self._screen.connected_printer)
+        logging.info("printer_cfg: %s", printer_cfg)
+        if printer_cfg is not None:
+            titlebar_items = printer_cfg.get("titlebar_items", "")
+            if titlebar_items is not None:
+                titlebar_items = [str(i.strip()) for i in titlebar_items.split(',')]
+                logging.info("Titlebar items: %s", titlebar_items)
+                self.titlebar_name_type = printer_cfg.get("titlebar_name_type", None)
+                logging.info("Titlebar name type: %s", self.titlebar_name_type)
+                for device in self._printer.get_temp_store_devices():
+                    # Users can fill the bar if they want
+                    if n >= nlimit + 1:
+                        break
+                    if not (device.startswith("extruder") or device.startswith("heater_bed")):
+                        name = device.split(" ")[1:][0]
+                    else:
+                        name = device
+                    for item in titlebar_items:
+                        if name == item:
+                            self.control['temp_box'].pack_start(self.labels["%s_box" % device], True, True, 3)
+                            n += 1
+                            break
+
+        # If there is enough space fill with heater_generic
+        for device in self._printer.get_temp_store_devices():
+            if n >= nlimit:
+                break
+            if device.startswith("heater_generic"):
+                self.control['temp_box'].pack_start(self.labels["%s_box" % device], True, True, 3)
+                n += 1
 
     def activate(self):
-        size = self.control['time_box'].get_allocation().width
-        self.layout.remove(self.control['time_box'])
-        self.control['time_box'].set_size_request(size, self.title_spacing)
-        self.layout.put(self.control['time_box'], self._screen.width - size - self.hmargin, 0)
-
-        GLib.timeout_add_seconds(1, self.update_time)
-        self.update_time()
+        if self.time_update is None:
+            self.time_update = GLib.timeout_add_seconds(1, self.update_time)
 
     def add_content(self, panel):
         self.current_panel = panel
@@ -201,17 +248,21 @@ class BasePanel(ScreenPanel):
         if action != "notify_status_update" or self._printer is None:
             return
 
-        if self._printer.has_heated_bed():
-            self.labels["heater_bed"].set_label(
-                "%02d°" % round(self._printer.get_dev_stat("heater_bed", "temperature")))
-        for x in self._printer.get_tools():
-            self.labels[x].set_label("%02d°" % round(self._printer.get_dev_stat(x, "temperature")))
+        for device in self._printer.get_temp_store_devices():
+            name = ""
+            if not (device.startswith("extruder") or device.startswith("heater_bed")):
+                if self.titlebar_name_type == "full":
+                    name = device.split(" ")[1:][0].capitalize().replace("_", " ") + ": "
+                elif self.titlebar_name_type == "short":
+                    name = device.split(" ")[1:][0][:1].upper() + ": "
+            self.labels[device].set_label("%s%d°" % (name, round(self._printer.get_dev_stat(device, "temperature"))))
 
         if "toolhead" in data and "extruder" in data["toolhead"]:
             if data["toolhead"]["extruder"] != self.current_extruder:
                 self.control['temp_box'].remove(self.labels["%s_box" % self.current_extruder])
                 self.current_extruder = data["toolhead"]["extruder"]
-                self.control['temp_box'].pack_start(self.labels["%s_box" % self.current_extruder], True, 3, 3)
+                self.control['temp_box'].pack_start(self.labels["%s_box" % self.current_extruder], True, True, 3)
+                self.control['temp_box'].reorder_child(self.labels["%s_box" % self.current_extruder], 0)
                 self.control['temp_box'].show_all()
 
 
@@ -325,7 +376,7 @@ class BasePanel(ScreenPanel):
         except Exception:
             logging.debug("Error parsing jinja for title: %s" % title)
 
-        self.titlelbl.set_label("%s | %s" % (self._screen.connected_printer, title))
+        self.titlelbl.set_label("%s | %s" % (self._screen.connecting_to_printer, title))
 
     def show_back_buttons(self):
         self.control_grid.attach(self.control['back'], 0, 0, 1, 1)
@@ -339,7 +390,7 @@ class BasePanel(ScreenPanel):
         confopt = self._config.get_main_config_option("24htime")
         if now.minute != self.time_min or self.time_format != confopt:
             if confopt == "True":
-                self.control['time'].set_text(now.strftime("     %H:%M"))
+                self.control['time'].set_text(now.strftime("%H:%M"))
             else:
                 self.control['time'].set_text(now.strftime("%I:%M %p"))
         return True