+++ panels/job_status.py
@@ -25,7 +25,7 @@ class JobStatusPanel(ScreenPanel):
     def initialize(self, panel_name):
         _ = self.lang.gettext
         self.timeleft_type = "file"
-        self.timeout = None
+        self.state_timeout = None
         self.close_timeouts = []
 
         self.create_buttons()
@@ -130,14 +130,21 @@ class JobStatusPanel(ScreenPanel):
         self.labels['elapsed'].get_style_context().add_class("printing-info")
         self.labels['duration'] = Gtk.Label(label="0s")
         self.labels['duration'].get_style_context().add_class("printing-info")
-        self.labels['est_time'] = Gtk.Label(label="/ 0s")
+        self.labels['total'] = Gtk.Label(label=_("Total:"))
+        self.labels['total'].get_style_context().add_class("printing-info")
+        self.labels['est_time'] = Gtk.Label(label="0s")
         self.labels['est_time'].get_style_context().add_class("printing-info")
-        it_box = Gtk.Box(spacing=0)
-        it_box.add(clock)
-        it_box.add(self.labels['elapsed'])
-        it_box.add(self.labels['duration'])
-        it_box.add(self.labels['est_time'])
-        self.labels['it_box'] = it_box
+        timegrid = Gtk.Grid()
+        it1_box = Gtk.Box(spacing=0)
+        it1_box.add(self.labels['elapsed'])
+        it1_box.add(self.labels['duration'])
+        it2_box = Gtk.Box(spacing=0)
+        it2_box.add(self.labels['total'])
+        it2_box.add(self.labels['est_time'])
+        timegrid.attach(clock, 0, 0, 1, 2)
+        timegrid.attach(it1_box, 1, 0, 1, 1)
+        timegrid.attach(it2_box, 1, 1, 1, 1)
+        self.labels['timegrid'] = timegrid
 
         position = self._gtk.Image("move.svg", None, .6, .6)
         self.labels['pos_x'] = Gtk.Label(label="X: 0")
@@ -228,8 +235,8 @@ class JobStatusPanel(ScreenPanel):
         _ = self.lang.gettext
         ps = self._printer.get_stat("print_stats")
         self.set_state(ps['state'])
-        if self.timeout is None:
-            GLib.timeout_add_seconds(1, self.state_check)
+        if self.state_timeout is None:
+            self.state_timeout = GLib.timeout_add_seconds(1, self.state_check)
 
     def add_labels(self):
         for child in self.labels['i1_box'].get_children():
@@ -241,7 +248,7 @@ class JobStatusPanel(ScreenPanel):
         self.labels['i2_box'].add(self.labels['temp_grid'])
         self.labels['i2_box'].add(self.labels['pos_box'])
         self.labels['i2_box'].add(self.labels['sfe_grid'])
-        self.labels['i2_box'].add(self.labels['it_box'])
+        self.labels['i2_box'].add(self.labels['timegrid'])
         self.labels['i2_box'].add(self.labels['itl_box'])
 
 
@@ -304,6 +311,7 @@ class JobStatusPanel(ScreenPanel):
             return
 
         logging.debug("Canceling print")
+        self.set_state("cancelling")
         self.disable_button("pause", "resume", "cancel")
         self._screen._ws.klippy.print_cancel(self._response_callback)
 
@@ -339,8 +347,8 @@ class JobStatusPanel(ScreenPanel):
 
     def new_print(self):
         self.remove_close_timeout()
-        if self.timeout is None:
-            GLib.timeout_add_seconds(1, self.state_check)
+        if self.state_timeout is None:
+            self.state_timeout = GLib.timeout_add_seconds(1, self.state_check)
         self._screen.wake_screen()
         self.state_check()
 
@@ -411,29 +419,57 @@ class JobStatusPanel(ScreenPanel):
             self.update_percent_complete()
 
         self.update_text("duration", str(self._gtk.formatTimeString(ps['print_duration'])))
+        self.update_text("time_left", self.calculate_time_left(ps['print_duration'], ps['filament_used']))
 
-        timeleft_type = self._config.get_config()['main'].get('print_estimate_method', 'file')
-        if timeleft_type != self.timeleft_type:
-            if self.timeleft_type == "duration":
-                self.labels['it_box'].add(self.labels['est_time'])
-            elif timeleft_type == "duration":
-                self.labels['it_box'].remove(self.labels['est_time'])
-            self.timeleft_type = timeleft_type
+    def calculate_time_left(self, duration=0, filament_used=0):
+        total_duration = None
+        if self.progress < 1:
+            slicer_time = filament_time = file_time = None
+            timeleft_type = self._config.get_config()['main'].get('print_estimate_method', 'auto')
+            slicer_correction = (self._config.get_config()['main'].getint('print_estimate_compensation', 100) / 100)
+            # speed_factor compensation based on empirical testing
+            spdcomp = math.sqrt(self.speed / 100)
 
-        if timeleft_type in ['filament', 'file', 'slicer']:
-            duration = ps['print_duration']
-            if timeleft_type == "filament":
-                estimated_filament = (self.file_metadata['filament_total'] if "filament_total" in self.file_metadata
-                                      else 1)
-                total_duration = duration / (max(ps['filament_used'], 0.0001) / max(estimated_filament, 0.0001))
-            elif timeleft_type == "file":
-                total_duration = duration / max(self.progress, 0.0001)
-            elif timeleft_type == "slicer":
-                total_duration = (self.file_metadata['estimated_time'] if "estimated_time" in self.file_metadata
-                                  else duration)
-            time_left = max(total_duration - duration, 0)
-            self.update_text("time_left", str(self._gtk.formatTimeString(time_left)))
-            self.update_text("est_time", "/ %s" % str(self._gtk.formatTimeString(total_duration)))
+            if "estimated_time" in self.file_metadata:
+                if self.file_metadata['estimated_time'] > 0:
+                    slicer_time = (self.file_metadata['estimated_time'] * slicer_correction) / spdcomp
+                    if slicer_time < duration:
+                        slicer_time = None
+
+            if "filament_total" in self.file_metadata:
+                if self.file_metadata['filament_total'] > 0 and filament_used > 0:
+                    if self.file_metadata['filament_total'] > filament_used:
+                        filament_time = duration / (filament_used / self.file_metadata['filament_total'])
+                        if filament_time < duration:
+                            filament_time = None
+
+            if self.progress > 0:
+                file_time = duration / self.progress
+
+            if timeleft_type == "file" and file_time is not None:
+                total_duration = file_time
+            elif timeleft_type == "filament" and filament_time is not None:
+                total_duration = filament_time
+            elif slicer_time is not None:
+                if timeleft_type == "slicer":
+                    total_duration = slicer_time
+                else:
+                    if filament_time is not None and self.progress > 0.14:
+                        # Weighted arithmetic mean (Slicer is the most accurate)
+                        total_duration = (slicer_time * 3 + filament_time + file_time) / 5
+                    else:
+                        # At the begining file and filament are innacurate
+                        total_duration = slicer_time
+            elif file_time is not None:
+                if filament_time is not None:
+                    total_duration = (filament_time + file_time) / 2
+                else:
+                    total_duration = file_time
+
+        if total_duration is None:
+            return "-"
+        self.update_text("est_time", str(self._gtk.formatTimeString(total_duration)))
+        return str(self._gtk.formatTimeString((total_duration - duration)))
 
     def state_check(self):
         ps = self._printer.get_stat("print_stats")
@@ -487,17 +523,17 @@ class JobStatusPanel(ScreenPanel):
 
         if self.state != state:
             logging.debug("Changing job_status state from '%s' to '%s'" % (self.state, state))
-        self.state = state
         if state == "paused":
             self.update_text("status", _("Paused"))
         elif state == "printing":
             self.update_text("status", _("Printing"))
         elif state == "cancelling":
             self.update_text("status", _("Cancelling"))
-        elif state == "cancelled":
+        elif state == "cancelled" or (state == "standby" and self.state == "cancelling"):
             self.update_text("status", _("Cancelled"))
         elif state == "complete":
             self.update_text("status", _("Complete"))
+        self.state = state
         self.show_buttons_for_state()
 
 
@@ -540,7 +576,7 @@ class JobStatusPanel(ScreenPanel):
             self.file_metadata = self._files.get_file_info(self.filename)
             logging.info("Update Metadata. File: %s Size: %s" % (self.filename, self.file_metadata['size']))
             if "estimated_time" in self.file_metadata and self.timeleft_type == "slicer":
-                self.update_text("est_time", "/ %s" %
+                self.update_text("est_time",
                                  str(self._gtk.formatTimeString(self.file_metadata['estimated_time'])))
             if "thumbnails" in self.file_metadata:
                 tmp = self.file_metadata['thumbnails'].copy()
@@ -566,7 +602,6 @@ class JobStatusPanel(ScreenPanel):
                         self.file_metadata['gcode_start_byte']))
         else:
             progress = self._printer.get_stat('virtual_sdcard', 'progress')
-        progress = round(progress, 2)
 
         if progress != self.progress:
             self.progress = progress